requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib
python-jose
boto3
